pyproject.toml
[tool.poetry]
name = "annotatedwastecomponents"
version = "0.1.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.2"
gunicorn = "^23.0.0"
uvicorn = "^0.31.1"
pandas = "^2.1.3"
setuptools = "^75.1.0"
pydantic = "^2.9.2"
gitpython = "^3.1.40"
loguru = "^0.7.1"
tqdm = "^4.66.1"
more-itertools = "^10.1.0"
sqlalchemy = "^2.0.21"
psycopg = {extras = ["binary"], version = "^3.1.14"}
streamlit = "^1.39.0"
streamlit-extras = "^0.5.0"
scipy = "^1.11.4"
wurlitzer = "^3.0.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.16.0"
hydra-core = "^1.3.2"
bayanpy = "^0.7.7"
cdlib = "^0.4.0"
networkit = "^11.0"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
openai = "^1.51.2"
langchain-openai = "^0.2.2"
leidenalg = "^0.10.2"
igraph = "^0.11.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
